Fixup libnetwork lint errors
Taken from https://github.com/moby/moby/commit/4b981436fe26724883cee71c15f218993ab53c5b

+++ bitseq/sequence_test.go
@@ -660,7 +660,7 @@ func TestOffsetSetUnset(t *testing.T) {
 		t.Fatal(err)
 	}
 
-	//At this point sequence is (0xffffffff, 9)->(0x7fffffff, 1)->(0xffffffff, 22)->end
+	// At this point sequence is (0xffffffff, 9)->(0x7fffffff, 1)->(0xffffffff, 22)->end
 	if o, err = hnd.SetAnyInRange(32, 500, false); err != nil {
 		t.Fatal(err)
 	}
@@ -847,7 +847,7 @@ func TestSetAnyInRange(t *testing.T) {
 
 func TestMethods(t *testing.T) {
 	numBits := uint64(256 * blockLen)
-	hnd, err := NewHandle("path/to/data", nil, "sequence1", uint64(numBits))
+	hnd, err := NewHandle("path/to/data", nil, "sequence1", numBits)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -1284,7 +1284,7 @@ func testSetRollover(t *testing.T, serial bool) {
 		t.Fatalf("Unexpected free bits: found %d free bits.\nSeed: %d.\n%s", hnd.unselected, seed, hnd)
 	}
 
-	//request to allocate for remaining half of the bits
+	// request to allocate for remaining half of the bits
 	for i := 0; i < numBits/2; i++ {
 		_, err := hnd.SetAny(serial)
 		if err != nil {
@@ -1292,8 +1292,8 @@ func testSetRollover(t *testing.T, serial bool) {
 		}
 	}
 
-	//At this point all the bits must be allocated except the randomly unallocated bits
-	//which were unallocated in the first half of the bit sequence
+	// At this point all the bits must be allocated except the randomly unallocated bits
+	// which were unallocated in the first half of the bit sequence
 	if hnd.Unselected() != uint64(numBits/4) {
 		t.Fatalf("Unexpected number of unselected bits %d, Expected %d", hnd.Unselected(), numBits/4)
 	}
@@ -1304,8 +1304,8 @@ func testSetRollover(t *testing.T, serial bool) {
 			t.Fatalf("Unexpected failure on allocation %d: %v\nSeed: %d\n%s", i, err, seed, hnd)
 		}
 	}
-	//Now requesting to allocate the unallocated random bits (qurter of the number of bits) should
-	//leave no more bits that can be allocated.
+	// Now requesting to allocate the unallocated random bits (qurter of the number of bits) should
+	// leave no more bits that can be allocated.
 	if hnd.Unselected() != 0 {
 		t.Fatalf("Unexpected number of unselected bits %d, Expected %d", hnd.Unselected(), 0)
 	}
